[BACKLOG-6739] Relocated will/did/rejected:change event classes and respective mixins.
