+++ package-res/resources/web/pentaho/type/_doc/_namespace.jsdoc
