+++ package-res/resources/web/pentaho/visual/base/_doc/_namespace.jsdoc
@@ -21,31 +21,3 @@
  * @name pentaho.visual.base
  * @namespace
  */
-
-
-/**
- * The `events` namespace contains event classes for developers to use
- * to build new visualizations that work with the Visualization API.
- *
- * @name pentaho.visual.base.events
- * @namespace
- *
- */
-
-/**
- * The `mixins` namespace contains mixins for developers to use
- * to build new visualizations that work with the Visualization API.
- *
- * @name pentaho.visual.base.mixins
- * @namespace
- *
- */
-
-/**
- * The `types` namespace contains data types for developers to use
- * to build new visualizations that work with the Visualization API.
- *
- * @name pentaho.visual.base.types
- * @namespace
- *
- */
